Updated imports/exports to clean errors and warnings
