+++ code/drasil-code/Language/Drasil/Code/Imperative/GenerateGOOL.hs
@@ -1,7 +1,20 @@
 module Language.Drasil.Code.Imperative.GenerateGOOL (
-  genModule, genDoxConfig, publicClass, publicMethod, publicInOutFunc, fApp, 
-  fAppInOut, value, variable, mkParam
+  genModule, genDoxConfig, publicClass, fApp, fAppInOut, mkParam
 ) where
+
+import Language.Drasil
+import Language.Drasil.Code.Code as C (CodeType(List, Object))
+import Language.Drasil.Code.Imperative.State (State(..))
+import Language.Drasil.Code.Imperative.GOOL.Symantics (Label, ProgramSym(..), 
+  RenderSym(..), AuxiliarySym(..), StateTypeSym(..), VariableSym(..), 
+  ValueSym(..), ValueExpression(..), StatementSym(..), ParameterSym(..),
+  MethodSym(..), StateVarSym(..), ClassSym(..), ModuleSym(..))
+import Language.Drasil.CodeSpec (CodeSpec(..), CodeSystInfo(..), Comments(..), 
+  Name)
+
+import qualified Data.Map as Map (lookup)
+import Data.Maybe (fromMaybe, maybe)
+import Control.Monad.Reader (Reader, ask, withReader)
 
 genModule :: (RenderSym repr) => Name -> String
   -> Maybe (Reader State [repr (Method repr)])
@@ -38,54 +51,6 @@ publicClass desc n l vs ms = do
     then docClass desc (pubClass n l vs ms) 
     else pubClass n l vs ms
 
-publicMethod :: (RenderSym repr, HasUID c, HasCodeType c, CodeIdea c) => 
-  repr (MethodType repr) -> Label -> String -> [c] -> Maybe String -> 
-  [repr (Block repr)] -> Reader State (repr (Method repr))
-publicMethod = genMethod public static_
-
-publicInOutFunc :: (RenderSym repr, HasUID c, HasCodeType c, CodeIdea c) => 
-  Label -> String -> [c] -> [c] -> [c] -> [repr (Block repr)] -> 
-  Reader State (repr (Method repr))
-publicInOutFunc = genInOutFunc public static_
-
-genMethod :: (RenderSym repr, HasUID c, HasCodeType c, CodeIdea c) => 
-  repr (Scope repr) -> repr (Permanence repr) -> repr (MethodType repr) -> 
-  Label -> String -> [c] -> Maybe String -> [repr (Block repr)] -> 
-  Reader State (repr (Method repr))
-genMethod s pr t n desc p r b = do
-  g <- ask
-  vars <- mapM mkVar p
-  let ps = map mkParam vars
-      doLog = logKind g
-      loggedBody LogFunc = loggedMethod (logName g) n vars b
-      loggedBody LogAll  = loggedMethod (logName g) n vars b
-      loggedBody _       = b
-      bod = body $ loggedBody doLog
-      fn = function n s pr t ps bod
-  pComms <- mapM (paramComment . (^. uid)) p
-  return $ if CommentFunc `elem` commented g
-    then docFunc desc pComms r fn else fn
-
-genInOutFunc :: (RenderSym repr, HasUID c, HasCodeType c, CodeIdea c) => 
-  repr (Scope repr) -> repr (Permanence repr) -> Label -> String -> [c] ->
-  [c] -> [c] -> [repr (Block repr)] -> Reader State (repr (Method repr))
-genInOutFunc s pr n desc ins outs both b = do
-  g <- ask
-  inVs <- mapM mkVar ins
-  outVs <- mapM mkVar outs
-  bothVs <- mapM mkVar both
-  let doLog = logKind g
-      loggedBody LogFunc = loggedMethod (logName g) n inVs b
-      loggedBody LogAll  = loggedMethod (logName g) n inVs b
-      loggedBody _       = b
-      bod = body $ loggedBody doLog
-      fn = inOutFunc n s pr inVs outVs bothVs bod
-  pComms <- mapM (paramComment . (^. uid)) ins
-  oComms <- mapM (paramComment . (^. uid)) outs
-  bComms <- mapM (paramComment . (^. uid)) both
-  return $ if CommentFunc `elem` commented g 
-    then docInOutFunc desc pComms oComms bComms fn else fn
-
 fApp :: (RenderSym repr) => String -> String -> repr (StateType repr) -> 
   [repr (Value repr)] -> Reader State (repr (Value repr))
 fApp m s t vl = do
@@ -99,31 +64,6 @@ fAppInOut m n ins outs both = do
   g <- ask
   return $ if m /= currentModule g then extInOutCall m n ins outs both
     else inOutCall n ins outs both
-
-value :: (RenderSym repr) => UID -> String -> repr (StateType repr) -> 
-  Reader State (repr (Value repr))
-value u s t = do
-  g <- ask
-  let cs = codeSpec g
-      mm = constMap cs
-  maybe (do { v <- variable s t; return $ valueOf v }) 
-    (convExpr . codeEquat) (Map.lookup u mm)
-
-variable :: (RenderSym repr) => String -> repr (StateType repr) -> 
-  Reader State (repr (Variable repr))
-variable s t = do
-  g <- ask
-  let cs = csi $ codeSpec g
-  if s `elem` map codeName (inputs cs) 
-    then inputVariable (inStruct g) s t
-    else return $ var s t
-  
-inputVariable :: (RenderSym repr) => Structure -> String -> 
-  repr (StateType repr) -> Reader State (repr (Variable repr))
-inputVariable Unbundled s t = return $ var s t
-inputVariable Bundled s t = do
-  ip <- mkVar (codevar inParams)
-  return $ ip $-> var s t
 
 mkParam :: (RenderSym repr) => repr (Variable repr) -> repr (Parameter repr)
 mkParam v = paramFunc (getType $ variableType v) v
