+++ code/drasil-code/Language/Drasil/Code/Imperative/Helpers.hs
@@ -1,17 +1,16 @@
 module Language.Drasil.Code.Imperative.Helpers (
-  liftS, mkVal, mkVar, getUpperBound, lookupC
+  liftS, getUpperBound, lookupC
 ) where
+
+import Language.Drasil
+import Database.Drasil (symbResolve)
+import Language.Drasil.Code.Imperative.State (State(..))
+import Language.Drasil.CodeSpec (CodeSpec(..), CodeSystInfo(..))
+
+import Control.Monad.Reader (Reader)
 
 liftS :: Reader a b -> Reader a [b]
 liftS = fmap (: [])
-
-mkVal :: (RenderSym repr, HasUID c, HasCodeType c, CodeIdea c) => c -> 
-  Reader State (repr (Value repr))
-mkVal v = value (v ^. uid) (codeName v) (convType $ codeType v)
-
-mkVar :: (RenderSym repr, HasCodeType c, CodeIdea c) => c -> 
-  Reader State (repr (Variable repr))
-mkVar v = variable (codeName v) (convType $ codeType v)
 
 getUpperBound :: Expr -> Expr
 getUpperBound (BinaryOp Lt _ b) = b
